# backend/swagger.yaml
openapi: 3.1.0
info:
  title: CoinCraft API
  description: |
    Financial literacy learning platform API for children, teens, parents, and teachers. 
    Each endpoint is mapped to specific user stories to ensure alignment with application requirements.
  version: 1.0.0
paths:
...
  /api/teachers/{teacher_id}/classes:
    get:
      tags:
      - Teachers
      summary: Get Teacher Classes
      description: Get all classes managed by a teacher.
      operationId: get_teacher_classes_api_teachers__teacher_id__classes_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: teacher_id
        in: path
        required: true
        schema:
          type: string
          title: Teacher Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassRead'
                title: Response Get Teacher Classes Api Teachers  Teacher Id  Classes
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Teachers
      summary: Create Class
      description: Create a new class for a teacher.
      operationId: create_class_api_teachers__teacher_id__classes_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: teacher_id
        in: path
        required: true
        schema:
          type: string
          title: Teacher Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/classes/{class_id}:
    put:
      tags:
      - Teachers
      summary: Update Class
      description: Update class information.
      operationId: update_class_api_classes__class_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Teachers
      summary: Delete Class
      description: Delete a class.
      operationId: delete_class_api_classes__class_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/classes/{class_id}/students:
    get:
      tags:
      - Teachers
      summary: Get Class Students
      description: Get all students in a class.
      operationId: get_class_students_api_classes__class_id__students_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentRead'
                title: Response Get Class Students Api Classes  Class Id  Students
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Teachers
      summary: Add Student To Class
      description: Add a student to a class.
      operationId: add_student_to_class_api_classes__class_id__students_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Student Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/classes/{class_id}/students/{student_id}:
    delete:
      tags:
      - Teachers
      summary: Remove Student From Class
      description: Remove a student from a class.
      operationId: remove_student_from_class_api_classes__class_id__students__student_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: class_id
        in: path
        required: true
        schema:
          type: string
          title: Class Id
      - name: student_id
        in: path
        required: true
        schema:
          type: string
          title: Student Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClassCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: ClassCreate
    ClassRead:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        id:
          type: string
          title: Id
        teacher_id:
          type: string
          title: Teacher Id
        class_code:
          type: string
          title: Class Code
        is_active:
          type: boolean
          title: Is Active
          default: true
        created_at:
          type: string
          format: date-time
          title: Created At
        students_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Students Count
          default: 0
        average_performance:
          anyOf:
          - type: number
          - type: 'null'
          title: Average Performance
          default: 0.0
      type: object
      required:
      - name
      - id
      - teacher_id
      - class_code
      - created_at
      title: ClassRead
    ClassUpdate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      required:
      - name
      title: ClassUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    StudentRead:
      properties:
        user_id:
          type: string
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        level:
          anyOf:
          - type: integer
          - type: 'null'
          title: Level
        performance_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Performance Score
        needs_support:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Needs Support
        last_activity_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Activity Date
        modules_completed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Modules Completed
        total_time_spent:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Time Spent
        progress:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Progress
      type: object
      required:
      - user_id
      - name
      - email
      title: StudentRead
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login